Make moderated grading radio buttons only show when needed
This makes it so that the radio buttons only get displayed
if there is at least two provisional grades to select from.
It also makes it so that if there is only one grade it shows
up in the Grade column at the end of the row indicating to
the user that that will be the grade selected unless something
changes.

fixes CNVS-27401

Test Plan:
  - Set up a moderated grading assignment
  - Grade the assignment for two students
  - Add a reviewer for one of the students
  - Grade that student with another provisional grade
  - On the moderation screen you should see that the
    student without a reviewer shows the grade, but no
    radio button.
  - The Grade column for this student should match the
    only grade that they have.
  - The student with two provisional grades should
    have a radio button to select between the grades.

Change-Id: Ice423550e96991f4a2f56b37d83be570aca08c14
Reviewed-on: https://gerrit.instructure.com/72810
Tested-by: Jenkins
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Product-Review: Cosme Salazar <cosme@instructure.com>