api: return an error if grade_student fails
grade_student has a number of validation checks it does an raises
GradeError if any of them fail.  this returns that error to the user.

The case that brought this up was trying to grade a concluded
enrollment. we allow looking up the user for
SubmissionsApiController#update if they are concluded, but not grading
them. Unfortunately the error message that gets returned in this case is
"Student must be enrolled in the course as a student to be graded",
which is a little confusing.  But in general this change is still
valuable to catch other things like trying to grade and excuse at the
same time.

Down the road we may want to restrict looking up concluded users, so
that we can return a more specific error message.

fixes CNVS-28732

test plan:
- hit the submission update endpoint with an invalid request (concluded
student or grade+excuse should both do it
- it should return a 400 and an error message

Change-Id: I45d2bf39f19fa07c76730dba1228513d8b667fd5
Reviewed-on: https://gerrit.instructure.com/77093
Tested-by: Jenkins
Reviewed-by: Shahbaz Javeed <sjaveed@instructure.com>
Reviewed-by: Jeremy Neander <jneander@instructure.com>
Reviewed-by: Keith T. Garner <kgarner@instructure.com>
QA-Review: Amber Taniuchi <amber@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>