apply body height to #tinyrce same as #tinymce
refs CNVS-26705

we make the body of the tinymce iframe taller so that (maybe among other
things) the drag/drop target in an empty or nearly empty editor is
larger. but the rule targeted #tinymce, and in the RCS (for reasons) the
body's id is #tinyrce. so make it work for both

test-plan:
 - enable remote editor
 - go to a page with an editor (e.g. create assignment)
 - make sure the editor contents are empty or nearly so
 - try and drag content into the editor with the mouse released over an empty
   portion of the editor (well below the end of any existing content)
 - should still drop the content into the editor, landing at the end of
   the existing content

Change-Id: I42d9760ee482449e6974f8adace2b7e10aa7d5dd
Reviewed-on: https://gerrit.instructure.com/77411
Reviewed-by: Ethan Vizitei <evizitei@instructure.com>
Tested-by: Jenkins
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>