a11y KO remove invisible link on eportfolio page
fixes CNVS-26678

When viewing an ePortfolio page, there is a list of pages in the
section that has an Edit This Page button below it. Previously,
between the last item in this list and the Edit This Page button,
there was an invisible link that points to the portfolio
dashboard/welcome page. Now the link has been hidden and will not
be displayed.

test plan
1. Navigate to the an ePortfolio
2. Tab to the last page listed on the right
3. Press Tab one more time to focus the invisible link
4. Notice how you cant find it.

Change-Id: I9177e5aa176a7dbb3a96c981464ef6b7ff1a16ac
Reviewed-on: https://gerrit.instructure.com/81678
Tested-by: Jenkins
QA-Review: Nathan Rogowski <nathan@instructure.com>
Reviewed-by: Matt Berns <mberns@instructure.com>
Product-Review: Pierce Arner <pierce@instructure.com>