use set_id when creating periods via batch update
previously, when a period was created as part of
the batch update process, it would always create
the period on the first grading period group
associated with the context. now, when a set_id
is passed to batch_update, grading periods are
created on the set with the matching set_id.

closes CNVS-30868

test plan:
 - enable multiple grading periods for an account
 - go to the account grading standards page and
   create at least 2 grading period sets
 - create at least one grading period in each of
   the grading period sets
 - refresh the page and verify the grading periods
   you created are contained within the correct
   grading period sets

Change-Id: I0f279496ac75af6240a3caf5259380cc7c51a402
Reviewed-on: https://gerrit.instructure.com/87284
Reviewed-by: Jeremy Neander <jneander@instructure.com>
Reviewed-by: Nicholas Pitrak <npitrak@instructure.com>
Tested-by: Jenkins
QA-Review: KC Naegle <knaegle@instructure.com>
Product-Review: Christi Wruck