Make JS_SPEC_MATCHER work again
After the karma upgrade some things changed around that essentially
made it not work at all.  This makes it so that it works again.

In testing when JS_SPEC_MATCHER was set to **/*BreadcrumbsSpec it matched
properly, but the tests failed for some reason.  Other tests seemed to
pass as expected.  I'm pretty sure it has something to dow with requirejs
not loading something properly, but it's pretty isolated..

Change-Id: I2720ba97252ee4d3c18f350c22a02aa58df93be3
Reviewed-on: https://gerrit.instructure.com/87681
Tested-by: Jenkins
Reviewed-by: Jeremy Neander <jneander@instructure.com>
Reviewed-by: Derek Bender <djbender@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Clay Diffrient <cdiffrient@instructure.com>