spec: fix sporadic failure
use a real session instead of a stubbed one. this spec would usually pass,
(like 98% of the time) but only because of the User#last_logged_out
failsafe... the session was never actually destroyed.

when it did fail, it was due to a subtle timing issue around second
boundaries and how we compare the session_refreshed_at and last_logged_out
times. the former is truncated to the second, whereas the latter is not.
last_logged_out gets set first, so if session_refreshed_at falls in the
next second and your session is otherwise valid, you are still considered
logged in, making the spec fail.

Change-Id: If50df7ddc7b7629233dc9ce7016d0990f46bedba
Reviewed-on: https://gerrit.instructure.com/89519
Tested-by: Jenkins
Reviewed-by: August Thornton <august@instructure.com>
Product-Review: Jon Jensen <jon@instructure.com>
QA-Review: Jon Jensen <jon@instructure.com>