fix incorrect .dockerignore of translations/_core_en.js
A docker ignore entry was recently added for:

   public/javascripts/translations/

This is a mirror of the .gitignore entry, however, there is one file
checked into the git repo that shouldn't be ignored, and is necessary
for docker image builds. This adds that file back.

Test Plan:
1. Docker image successfully builds from running:
   docker-compose build web

Change-Id: Ia83fb6673197cb5cb8f98d9c230d65342b6be37c
Reviewed-on: https://gerrit.instructure.com/97721
Reviewed-by: Ryan Norton <rnorton@instructure.com>
Tested-by: Jenkins
Product-Review: Bryan Petty <bpetty@instructure.com>
QA-Review: Bryan Petty <bpetty@instructure.com>