pop-up event in calendar for appointment group reserved slots.
- also redirect student to find appointment mode if they
  do not have appointment.
- add additional specs, remove parameter to specify
  find_appointment statically.
- add specs for group appointments.

Test Plan:
Verify that the new appointment available notification
correctly routes the user to the appropriate calendar view.
When the user has reserved an appointment, the user should
be directed to the agenda view on the day of their appointment.
When the user does not have an appointment, find_appointment
mode should be activated in the agenda view of the calendar.
Links to the event in the syllabus should also correctly route
the user to the correct find_appointment mode based on whether
or not they have an appointment.

Additionally the  “Appointment Reserved by User”
notification should be tested.  Proposed change can be tested by
creating an appointment group as a teacher, and creating an event
with several slots for students.  As a student, sign up for a slot.
The teacher should receive a notification.  As the teacher,
follow the link in the notification, and verify that the
appointment event opens as a pop-up in the calendar month view.

Fixes PFS-6257, PFS-6318

Change-Id: Id906d77813bd2bf7fd3f4c3b9ca736accafe0cf2
Reviewed-on: https://gerrit.instructure.com/98922
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: David Tan <dtan@instructure.com>
Tested-by: Jenkins
Product-Review: Chris Ward <cward@instructure.com>