resize media comments dialog to fit video
When audio or video recordings are used in submission comments on
an assignment (in SpeedGrader or Submission Details), clicking
the thumbnail opens a modal that automatically plays the media,
but displays no controls, making it impossible to pause or
rewind.

This commit resizes the dialog so the controls are not hidden in
the overflow of the element.

fixes CNVS-30875

test plan:
* create a media comment
* click the thumbnail for the media comment
* verify the controls are visible

Change-Id: I8d1eadccec28c42c957f95546af3af762b2231dd
Reviewed-on: https://gerrit.instructure.com/100551
Reviewed-by: Shahbaz Javeed <sjaveed@instructure.com>
Reviewed-by: Derek Bender <djbender@instructure.com>
Tested-by: Jenkins
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>