more fixes to prepare for strict mode
refs: CNVS-35872

this fixes some eslint errors in order to get us towards strict mode.

to find these problems I enabled babel and ran the qUnit specs and
fixed the errors that it had.

Then I enabled an eslint config (g/106222) that only checked for 
possible runtime errors (in this case, mostly function statements 
that were inside an if/else and not the root of a function body 
which is not ok in strict mode) and did nothing about our code-style 
stuff. I uploaded patchset 1 on top of that commit so you can see
in the gergich comments that there is no errors when just checking
those things.

test plan:
* specs should pass


Change-Id: Id23c5b6a13c6366b7e8b6b298a9dee41a6950d4f
Reviewed-on: https://gerrit.instructure.com/106300
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>