handle equation and links plugin via event for rcs
this matches the pattern for other tinymce plugins not migrated to
the rcs. these plugins require jquery-ui and distributing it with the
module causes conflicts in canvas. they will be rewritten with react
when they are brought back in the the module.

refs CNVS-36075

test plan:
- the link and equation plugins for tinymce work the same as before
- test without rcs enabled as it is currently distributing its own
  version of these plugins

Change-Id: Iea86e7c82c06efa6984e815cf86c89478a6ed12a
Reviewed-on: https://gerrit.instructure.com/107337
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Brent Burgoyne <bburgoyne@instructure.com>
Tested-by: Jenkins