spec: fix failing chrome gradezilla spec
spec/selenium/grades/gradezilla/gradebook_spec.rb:226

the error was showing that the grid underneath will be clicked
instead of the download link. this might be due to the link being
off the viewable portion of the page. the attempted solution
is to make the page full screen so that the link is visible

Change-Id: Icb493bfb5ce2d6c5a8ee0a79d465a245a8144b03
Reviewed-on: https://gerrit.instructure.com/108527
Tested-by: Jenkins
Reviewed-by: Matt Fairbourn <mfairbourn@instructure.com>
Product-Review: Matt Fairbourn <mfairbourn@instructure.com>
QA-Review: Matt Fairbourn <mfairbourn@instructure.com>