conversation msg modals focus close button on open
For a11y, the compose and reply modals for conversations correctly focus
the close button when the dialog opens instead of the first control.

There are no unit test changes because I'm removing code for the old
explicit behavior, which wasn't tested. Now it has the default behavior
for modals, which is functionality that should be tested with the modal
modal implementation we use, and not with this particular use of it.

fixes CNVS-36820

test plan:
* open the conversations compose modal
* see the focus is set to the close button of the modal
* ditto with the reply modal

Change-Id: I25d3ae85a597e93ca2a2b00e240dee250adfbe40
Reviewed-on: https://gerrit.instructure.com/111573
Tested-by: Jenkins
Reviewed-by: Felix Milea-Ciobanu <fmileaciobanu@instructure.com>
QA-Review: David Tan <dtan@instructure.com>
Product-Review: Jon Willesen <jonw+gerrit@instructure.com>