Rename docker-compose.test.yml
We need to do this because docker hub tries to run the tests when
test.yml is present and our current configuration won't allow us to
disable it. As a stopgap we're working around the sadness by moving the
file to a new name.

Change-Id: I09ce965da07f1207b58fe589d03599aa4200c242
Reviewed-on: https://gerrit.instructure.com/112101
Tested-by: Jenkins
Reviewed-by: Jon Jensen <jon@instructure.com>
Product-Review: Jon Jensen <jon@instructure.com>
QA-Review: Jon Jensen <jon@instructure.com>