Make it so you can use js in app/coffescripts
closes:  CNVS-37356

in order to help the transition away from coffeescript, this makes
it so you can just change the file extension of a file in 
app/coffeescripts from .coffee to .js and convert it to js syntax.
the big win is that you won't have to change the 'require' path to it
in all the modules that were consuming it.

test plan:

* check out g/114236, it converts a bunch of files to .js and uses this
* yarn test and yarn run webpack should work

Change-Id: Ie4e22153c419bb92ebef66fb8dae196db08ff583
Reviewed-on: https://gerrit.instructure.com/114234
Tested-by: Jenkins
Reviewed-by: brian kirkby <bkirkby@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>