Replace forward slashes in filenames sent from turnitin with dash
Fixes: PLAT-2623

If a student submits to the Turnitin LTI via Google Drive and the filename
in Google Drive contains a forward slash, the submission will not be
returned to Canvas.

This changes the behavior so that we replace the filename with a dash
before we try to do anything else with it.

Test Plan:
- Install the TII LTI in your account
- Create a course with a TII LTI assignment
- Create a google doc with a forward-slash
- As a teacher, launch the TII assignment to configure it
- As a student, launch the TII Assignment
- Click on "Upload submisson" --> Additional Options --> google drive
- Authenticate with google drive and choose the google doc your previously
create
- After some time, the submission should be returned to canvas with a dash
in place of the forward slash

Change-Id: I56ea8be56d0d7a2671cd42dd8e976cec832016b6
Reviewed-on: https://gerrit.instructure.com/114470
Tested-by: Jenkins
Reviewed-by: Nathan Mills <nathanm@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>