Include inactive users in group members.
In the groups index page for a course, we were not displaying
inactive users.  This caused problems when a student tried to
sign up for a full group that had inactive users in it, since
we told them that the group had spaces.

This change now displays inactive as well as active users, so
the student is now told that the group is full.  It is up to
the teacher to remove such students if they see fit.

Fixes COMMS-4

Test plan:
* Have a course with some self sign-up groups, with a limit.
* Fill the group up to the limit with active students, making
  sure you have a student left that is not in the group.
* Now, deactivate one of the students in the group.
* With the student *not* in the group, act as that user.
* Go to the "people" page for the course and click on groups.
* Now the student will be told that the group is full.
* Become the teacher again.
* Go to the group in question, the inactive users in it should
  be marked as such.

Change-Id: If54df1e22779a5c0a4a97ed79245e7d5a6d76713
Reviewed-on: https://gerrit.instructure.com/131728
Tested-by: Jenkins
Reviewed-by: Felix Milea-Ciobanu <fmileaciobanu@instructure.com>
QA-Review: KC Naegle <knaegle@instructure.com>
Product-Review: Matt Goodwin <mattg@instructure.com>