update submissions using all assignments
fixes GRADE-375

QA Note:

Verifying correct behavior for this will be difficult. This is a
race condition that cannot be verified locally, due to how local
development works. This will likely require a portal.

The only way to verify that the conditions which produced the
original bug are present is to monitor the AJAX requests made by
the browser.  When the request(s) for submissions have completed
before one or more requests for assignment groups, the bug
behavior will be present without the changes from this commit.

test plan:
 1. Select or create:
    a. a Course
    b. a lot of Assignments
    c. 1 Student
 2. Visit SRGB
 3. Ensure the bug conditions are met (see above)
 4. Select the student.
 5. Verify the assigned assignments are listed in the Select
    Assignment menu
 6. Verify that the student can be graded on any given assignment

Change-Id: If93bf49e783ed090b1c36d9f82b8834caa8bdf30
Reviewed-on: https://gerrit.instructure.com/131793
Tested-by: Jenkins
Reviewed-by: Keith T. Garner <kgarner@instructure.com>
Reviewed-by: Derek Bender <djbender@instructure.com>
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>